Split up test_ode.py
