+++ pymc/tests/ode/test_ode.py
@@ -18,52 +18,12 @@ import aesara.tensor as at
 import numpy as np
 import pytest
 
-from scipy.integrate import odeint
 from scipy.stats import norm
 
 import pymc as pm
 
 from pymc.ode import DifferentialEquation
-from pymc.ode.utils import augment_system
 from pymc.tests.helpers import fast_unstable_sampling_mode
-
-
-def test_gradients():
-    """Tests the computation of the sensitivities from the Aesara computation graph"""
-
-    # ODE system for which to compute gradients
-    def ode_func(y, t, p):
-        return np.exp(-t) - p[0] * y[0]
-
-    # Computation of graidients with Aesara
-    augmented_ode_func = augment_system(ode_func, 1, 1 + 1)
-
-    # This is the new system, ODE + Sensitivities, which will be integrated
-    def augmented_system(Y, t, p):
-        dydt, ddt_dydp = augmented_ode_func(Y[:1], t, p, Y[1:])
-        derivatives = np.concatenate([dydt, ddt_dydp])
-        return derivatives
-
-    # Create real sensitivities
-    y0 = 0.0
-    t = np.arange(0, 12, 0.25).reshape(-1, 1)
-    a = 0.472
-    p = np.array([y0, a])
-
-    # Derivatives of the analytic solution with respect to y0 and alpha
-    # Treat y0 like a parameter and solve analytically.  Then differentiate.
-    # I used CAS to get these derivatives
-    y0_sensitivity = np.exp(-a * t)
-    a_sensitivity = (
-        -(np.exp(t * (a - 1)) - 1 + (a - 1) * (y0 * a - y0 - 1) * t) * np.exp(-a * t) / (a - 1) ** 2
-    )
-
-    sensitivity = np.c_[y0_sensitivity, a_sensitivity]
-
-    integrated_solutions = odeint(func=augmented_system, y0=[y0, 1, 0], t=t.ravel(), args=(p,))
-    simulated_sensitivity = integrated_solutions[:, 1:]
-
-    np.testing.assert_allclose(sensitivity, simulated_sensitivity, rtol=1e-5)
 
 
 def test_simulate():
